virtio-vsock: move shared types to lib.rs
Move PacketHeader, Error, Result, and PKT_HEADER_SIZE from the packet
module into lib.rs so they can be shared across the crate without
circular dependencies.
